chore(lint): add eslint, fix lint errors
Resolves #1.
