+++ index.js
@@ -1,9 +1,7 @@
 #!/usr/bin/env node
 
-'use strict';
-
 require('yargs')
-    .commandDir('cmds')
-    .demandCommand()
-    .help()
-    .argv;
+  .commandDir('lib/cmds')
+  .demandCommand()
+  .help()
+  .parse();
