+++ lib/util/logger.js
@@ -1,0 +1,7 @@
+/* eslint-disable no-console */
+module.exports = {
+  info: msg => console.info(`INFO: ${msg}`),
+  warn: msg => console.warn(`WARN: ${msg}`),
+  error: msg => console.error(`ERROR: ${msg}`),
+};
+/* eslint-enable no-console */
