+++ package.json
@@ -3,7 +3,9 @@
   "version": "0.0.1",
   "description": "A simple CLI to work with the Brotli compression algorithm",
   "main": "index.js",
-  "scripts": {},
+  "scripts": {
+    "lint": "eslint **/*.js"
+  },
   "bin": {
     "brotli-cli": "bin/brotli-cli.js"
   },
@@ -21,5 +23,10 @@
   "dependencies": {
     "brotli": "^1.3.2",
     "yargs": "^9.0.1"
+  },
+  "devDependencies": {
+    "eslint": "^4.9.0",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-plugin-import": "^2.8.0"
   }
 }
